New Midterm project added
A new Midterm project was made to keep changes to the code separate. This project contains the most up to date changes. At the moment, the program passes all the preliminary tests and addresses all the major requirements for the assignment. 

More tests should be made. Additionally, the code should be streamlined and cleaned where possible.