Added Final via upload
This is the most current version of the project. The most recent change was the addition to continue the game for as long as the user chose. There were also some minor changes such as the winning message, the addition of a "inProgress" instance variable and some formatting of the code for increased readability.